requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
flasgger==0.9.5
Flask==2.2.3
Flask-Cors==3.0.9
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.13.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
mistune==2.0.5
mongoengine==0.27.0
packaging==23.0
PyJWT==2.6.0
pymongo==4.3.3
pyrsistent==0.19.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
Unidecode==1.3.6
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
xmltodict==0.13.0
